% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_aois.R
\name{extract_aois}
\alias{extract_aois}
\title{Extract AOI-related Columns from Webcam Files and Calculate Locations}
\usage{
extract_aois(file_paths, zone_names = NULL)
}
\arguments{
\item{file_paths}{A list of file paths to webcam files (in .xlsx format).}
}
\value{
A dataframe containing distinct rows with AOI-related columns and calculated coordinates.
}
\description{
This function reads in multiple Gorilla webcam files, extracts the
`loc`, `x_normalised`, `y_normalised`, `width_normalised`,
and `height_normalised` columns, and calculates the bounding box coordinates
for the AOIs. It also rounds all numeric columns to 3 decimal places.
}
