% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_location.R
\name{find_location}
\alias{find_location}
\title{Find Image Location in a Given Set of Locations}
\usage{
find_location(locations, image)
}
\arguments{
\item{locations}{A character vector of image names at each location.}

\item{image}{A character string: the image to locate.#'}
}
\value{
A character string representing the location name or NA if not found.
}
\description{
Find Image Location in a Given Set of Locations
}
