% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_gaze.R
\name{smooth_gaze}
\alias{smooth_gaze}
\title{Apply a moving average smoothing function to gaze data (X and Y). This is generally reccomended after up-sampling the data.}
\usage{
smooth_gaze(x, n = 5, x_col = "Gaze_X", y_col = "Gaze_Y", trial_col = "Trial")
}
\arguments{
\item{x}{A data frame containing gaze data.}

\item{n}{The window size (in samples) for the moving average.}

\item{x_col}{The name of the X gaze column (as string).}

\item{y_col}{The name of the Y gaze column (as string).}

\item{trial_col}{The name of the trial column used for grouping (default = "Trial").}
}
\value{
A tibble with smoothed gaze X and Y columns (replacing originals).
}
\description{
Apply a moving average smoothing function to gaze data (X and Y). This is generally reccomended after up-sampling the data.
}
