% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample_gaze.R
\name{upsample_gaze}
\alias{upsample_gaze}
\title{Up-sample gaze and pupil data}
\usage{
upsample_gaze(
  x,
  pupil_cols = c("Pupil_Diameter"),
  gaze_cols = c("x_pred_normalised", "y_pred_normalised"),
  target_hz = 1000,
  upsample_pupil = TRUE
)
}
\arguments{
\item{x}{A dataframe containing gaze and pupil data with columns: `subject`, `trial`, and `time`.}

\item{pupil_cols}{Character vector of pupil diameter column names.}

\item{gaze_cols}{Character vector of gaze position column names.}

\item{target_hz}{Target sampling frequency (default is 1000 Hz).}

\item{upsample_pupil}{Logical; if `TRUE`, pupil data will also be upsampled.}
}
\value{
A dataframe with up-sampled time points and an `up_sampled` column.
}
\description{
Increase the sampling frequency to `target_hz` Hz by inserting additional rows.
Missing values in gaze and pupil data will be preserved for later interpolation.
}
