% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webgazeR-package.R
\docType{package}
\name{webgazeR-package}
\alias{webgazeR}
\alias{webgazeR-package}
\title{webgazeR: Tools for Processing Webcam Eye Tracking Data}
\description{
A companion package to gazeR. Functions for reading and pre-processing webcam eye tracking data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jgeller112.github.io/webgazeR/}
  \item \url{https://github.com/jgeller112/webgazeR}
}

}
\author{
\strong{Maintainer}: Jason Geller \email{drjasongeller@gmail.edu}

}
\keyword{internal}
