% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_gaze.R
\name{downsample_gaze}
\alias{downsample_gaze}
\title{Downsample gaze data}
\usage{
downsample_gaze(
  dataframe,
  bin.length = 50,
  timevar = "time",
  aggvars = c("subject", "condition", "target", "trial", "object", "time_bin")
)
}
\arguments{
\item{dataframe}{DataFrame containing gaze data.}

\item{bin.length}{Length of time bins (in milliseconds).}

\item{timevar}{Column name representing time.}

\item{aggvars}{Vector of variable names to group by for aggregation. Use "none" to skip aggregation.}
}
\value{
DataFrame with time bins added and optionally aggregated data.
}
\description{
This function combines gaze samples into time bins and optionally aggregates the data.
}
