% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sampling_rate.R
\name{filter_sampling_rate}
\alias{filter_sampling_rate}
\title{Filter or Label Data Based on Sampling Rate Threshold}
\usage{
filter_sampling_rate(
  data,
  threshold = NA,
  action = c("remove", "label"),
  by = c("subject", "trial", "both")
)
}
\arguments{
\item{data}{A dataframe with columns: subject, trial, SR_subject, SR_trial.}

\item{threshold}{Numeric. Sampling rate threshold to apply.}

\item{action}{"remove" (default) to delete bad data or "label" to flag bad data.}

\item{by}{"subject", "trial", or "both" to specify where to apply the threshold.}
}
\value{
A dataframe with either rows removed or bad subjects/trials labeled.
}
\description{
This function applies a sampling rate threshold and either removes
or labels "bad" subjects/trials based on their sampling rates.
}
