% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_location.R
\name{find_location}
\alias{find_location}
\title{Find Image Location in a Given Set of Locations}
\usage{
find_location(locations, image)
}
\arguments{
\item{locations}{A character vector representing the possible locations (e.g., `c("TL", "TR", "BL", "BR")`).}

\item{image}{A character value representing the image to find in the locations.}
}
\value{
A character string representing the location of the image, or `NA` if the image is not found or is missing.
}
\description{
This function determines the location of an image within a set of locations.
The function accepts a vector of locations (such as "TL", "TR", "BL", "BR") and an image identifier.
It returns the corresponding location name if the image is found, or `NA` if the image is not present or is `NA`.
}
\examples{
# Example usage of the find_location function
locations <- c("apple", "banana", "cherry", "date")
find_location(locations, "banana")  # Returns "TR" if locations follow c("TL", "TR", "BL", "BR")
find_location(locations, "orange")  # Returns NA

}
